+++ Helper/Webhook.php
@@ -3,6 +3,7 @@
 namespace Klaviyo\Reclaim\Helper;
 
 use Exception;
+use \Klaviyo\Reclaim\Helper\Logger;
 use \Klaviyo\Reclaim\Helper\ScopeSetting;
 
 class Webhook extends \Magento\Framework\App\Helper\AbstractHelper
@@ -11,14 +12,22 @@ class Webhook extends \Magento\Framework\App\Helper\AbstractHelper
     const WEBHOOK_URL = 'https://www.klaviyo.com/api/webhook/integration/magento_two';
 
     /**
+     * Klaviyo logger helper
+     * @var \Klaviyo\Reclaim\Helper\Logger $klaviyoLogger
+     */
+    protected $_klaviyoLogger;
+
+    /**
      * Klaviyo scope setting helper
      * @var \Klaviyo\Reclaim\Helper\ScopeSetting $klaviyoScopeSetting
      */
     protected $_klaviyoScopeSetting;
 
     public function __construct(
+        Logger $klaviyoLogger,
         ScopeSetting $klaviyoScopeSetting
     ) {
+        $this->_klaviyoLogger = $klaviyoLogger;
         $this->_klaviyoScopeSetting = $klaviyoScopeSetting;
     }
 
@@ -31,13 +40,13 @@ class Webhook extends \Magento\Framework\App\Helper\AbstractHelper
      */
     public function makeWebhookRequest($webhookType, $data, $klaviyoId=null)
     {
+
         if (!$klaviyoId) {
             $klaviyoId = $this->_klaviyoScopeSetting->getPublicApiKey();
         }
         $url = self::WEBHOOK_URL . '?c=' . $klaviyoId;
 
         $curl = curl_init();
-
         curl_setopt_array($curl, [
             CURLOPT_URL => $url,
             CURLOPT_RETURNTRANSFER => true,
@@ -57,7 +66,7 @@ class Webhook extends \Magento\Framework\App\Helper\AbstractHelper
         $err = curl_errno($curl);
 
         if ($err) {
-            throw new Exception(curl_error($curl));
+            $this->_klaviyoLogger->log(sprintf('Unable to send webhook to %s with data: %s', $url, json_encode($data)));
         }
 
         // Close cURL session handle