+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### [Unreleased]
+
+#### Added
 - Added a.fast.klaviyo.com to img-src csp whitelist
+
+#### Changed
 - Removed csp mode from config.xml
 
 ### [1.2.4] - 2020-12-01